Fix calculation of potentailly overlapping points
The current implementation can fail with a ValueError: math domain error raised in line 219, because (1-(xi-x)**2) can become negative. This is due to a missing call to abs(), which is there in the R code: https://github.com/aroneklund/beeswarm/blob/master/R/beeswarm.R#L348

With the call to abs() in place, the ValueError does not occur anymore.